Move EvidenceError to evidence.
